backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
cryptography
python-jose[cryptography]
passlib[bcrypt]
python-multipart
cloudinary
google-generativeai
python-dotenv
pydantic[email]
httpx